pyproject.toml
[tool.poetry]
name = "giobot"
version = "0.1.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.7"
httpx = "^0.12.1"
firebase-admin = "^4.3.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"